requirements.txt
mysqlclient
schedule 
requests
tinkoff-investments
python-dateutil